requirements.txt
beir
rank-bm25
sentence-transformers
faiss-cpu
tqdm
numpy
